# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CDN Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (superadmin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
  /users/{userId}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user role (superadmin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
        "404":
          description: User not found
...
  /domains/{domainId}/origins:
    get:
      operationId: listOrigins
      tags: [origins]
      summary: List origin servers for a domain
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of origin servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Origin"
    post:
      operationId: createOrigin
      tags: [origins]
      summary: Add an origin server to a domain
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOriginBody"
      responses:
        "201":
          description: Origin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Origin"
        "400":
          description: Validation error
        "403":
          description: Forbidden
  /domains/{domainId}/origins/{originId}:
    patch:
      operationId: updateOrigin
      tags: [origins]
      summary: Update an origin server
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: integer
        - name: originId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOriginBody"
      responses:
        "200":
          description: Updated origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Origin"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteOrigin
      tags: [origins]
      summary: Delete an origin server
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: integer
        - name: originId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /domains/{domainId}/stats:
    get:
      operationId: getDomainStats
      tags: [stats]
      summary: Get stats for a specific domain
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Domain stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainStats"
        "404":
          description: Not found
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: string
          nullable: true
        role:
          type: string
          enum: [superadmin, admin, user]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - role
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [superadmin, admin, user]
      required:
        - role
...
    Origin:
      type: object
      properties:
        id:
          type: integer
        domainId:
          type: integer
        address:
          type: string
          description: IP address or hostname of the origin server
        port:
          type: integer
        protocol:
          type: string
          enum: [http, https]
        weight:
          type: integer
          description: Load balancing weight (1-100)
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - domainId
        - address
        - port
        - protocol
        - weight
        - isActive
        - createdAt
    CreateOriginBody:
      type: object
      properties:
        address:
          type: string
        port:
          type: integer
        protocol:
          type: string
          enum: [http, https]
        weight:
          type: integer
      required:
        - address
    UpdateOriginBody:
      type: object
      properties:
        address:
          type: string
        port:
          type: integer
        protocol:
          type: string
          enum: [http, https]
        weight:
          type: integer
        isActive:
          type: boolean
...
    DomainStats:
      type: object
      properties:
        domainId:
          type: integer
        requests:
          type: integer
        cacheHits:
          type: integer
        cacheMisses:
          type: integer
        cacheHitRate:
          type: number
        bandwidth:
          type: integer
        cachedFiles:
          type: integer
        cacheSize:
          type: integer
          description: Cache size in bytes
      required:
        - domainId
        - requests
        - cacheHits
        - cacheMisses
        - cacheHitRate
        - bandwidth
        - cachedFiles
        - cacheSize