# lib/api-zod/src/generated/types/createOriginBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CDN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOriginBodyProtocol } from "./createOriginBodyProtocol";

export interface CreateOriginBody {
  address: string;
  port?: number;
  protocol?: CreateOriginBodyProtocol;
  weight?: number;
}