# lib/api-zod/src/generated/types/origin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CDN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { OriginProtocol } from "./originProtocol";

export interface Origin {
  id: number;
  domainId: number;
  /** IP address or hostname of the origin server */
  address: string;
  port: number;
  protocol: OriginProtocol;
  /** Load balancing weight (1-100) */
  weight: number;
  isActive: boolean;
  createdAt: Date;
}