# lib/api-zod/src/generated/types/updateOriginBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CDN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateOriginBodyProtocol } from "./updateOriginBodyProtocol";

export interface UpdateOriginBody {
  address?: string;
  port?: number;
  protocol?: UpdateOriginBodyProtocol;
  weight?: number;
  isActive?: boolean;
}